Process keywords in a predefined order
Since a keyword may defined in multiple chapters, e.g. ADD which is
defined in chapters 6.3, 7.3, 8.3, 9.3, and 10.3, it is important to
always link to the same chapter. To achieve this, we will always link
to the first occuring definition. "First" here means to the chapter
we the smalles numerical value. So ADD would here link to the definition
in chapter 6.3 since 6.3 is smaller than 7.3, 8.3, 9.3 and 10.3.